requirements.txt
numpy
pandas
dotenv

rasterio
geopandas
shapely

matplotlib
seaborn

torch>=2.0.0
torchvision
torchmetrics
torchsummary

affine
Pillow

scikit-learn

python-dotenv
tqdm

requests
fastapi
prefect
uvicorn
gensim

streamlit
streamlit_image_comparison
streamlit_folium
folium
albumentations
ultralytics

ipywidgets
